+++ mapper/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.7.8
+
+* #108, #109, feat: Flags `JsonProperty.notNull` `JsonProperty.required` spanned over serialization / deserialization process consistently
+
 ## 1.7.7
 
 * #106, fix: Do not generate invalid code when there is no main entry point provided, in case of library package for instance.